Split magick worker test from the main test.
